Use double matrices and vectors instead of float

+++ include/homogeneous.h
@@ -1,11 +1,11 @@
 #include <Eigen/Core>
 
-Eigen::MatrixXf toHomogeneous(const Eigen::MatrixXf & mat)
+Eigen::MatrixXd toHomogeneous(const Eigen::MatrixXd & mat)
 {
   const auto rows = mat.rows();
   const auto cols = mat.cols();
 
-  Eigen::MatrixXf out(rows + 1, cols);
+  Eigen::MatrixXd out(rows + 1, cols);
   out.block(0, 0, rows, cols) = mat;
   for (int c = 0; c < cols; c++) {
     out(rows, c) = 1.0;
